feat: service-break-06-001 - Compose domain services through the initializer
